+++ crates/transaction-pool/src/validate/eth.rs
@@ -21,10 +21,7 @@ use reth_primitives::{
 };
 use reth_storage_api::{AccountReader, StateProviderFactory};
 use reth_tasks::TaskSpawner;
-use revm::{
-    interpreter::gas::validate_initial_tx_gas,
-    primitives::{EnvKzgSettings, SpecId},
-};
+use revm::primitives::EnvKzgSettings;
 use std::{
     marker::PhantomData,
     sync::{atomic::AtomicBool, Arc},
@@ -287,7 +284,7 @@ where
             }
         }
 
-        if let Err(err) = ensure_intrinsic_gas(&transaction, &self.fork_tracker) {
+        if let Err(err) = transaction.ensure_intrinsic_gas(&self.fork_tracker) {
             return TransactionValidationOutcome::Invalid(transaction, err)
         }
 
@@ -783,35 +780,35 @@ impl ForkTracker {
     }
 }
 
-/// Ensures that gas limit of the transaction exceeds the intrinsic gas of the transaction.
-///
-/// Caution: This only checks past the Merge hardfork.
-pub fn ensure_intrinsic_gas<T: EthPoolTransaction>(
-    transaction: &T,
-    fork_tracker: &ForkTracker,
-) -> Result<(), InvalidPoolTransactionError> {
-    let spec_id = if fork_tracker.is_prague_activated() {
-        SpecId::PRAGUE
-    } else if fork_tracker.is_shanghai_activated() {
-        SpecId::SHANGHAI
-    } else {
-        SpecId::MERGE
-    };
+// /// Ensures that gas limit of the transaction exceeds the intrinsic gas of the transaction.
+// ///
+// /// Caution: This only checks past the Merge hardfork.
+// pub fn ensure_intrinsic_gas<T: EthPoolTransaction>(
+//     transaction: &T,
+//     fork_tracker: &ForkTracker,
+// ) -> Result<(), InvalidPoolTransactionError> {
+//     let spec_id = if fork_tracker.is_prague_activated() {
+//         SpecId::PRAGUE
+//     } else if fork_tracker.is_shanghai_activated() {
+//         SpecId::SHANGHAI
+//     } else {
+//         SpecId::MERGE
+//     };
 
-    let gas_after_merge = validate_initial_tx_gas(
-        spec_id,
-        transaction.input(),
-        transaction.kind().is_create(),
-        transaction.access_list().map(|list| list.0.as_slice()).unwrap_or(&[]),
-        transaction.authorization_count() as u64,
-    );
+//     let gas_after_merge = validate_initial_tx_gas(
+//         spec_id,
+//         transaction.input(),
+//         transaction.kind().is_create(),
+//         transaction.access_list().map(|list| list.0.as_slice()).unwrap_or(&[]),
+//         transaction.authorization_count() as u64,
+//     );
 
-    if transaction.gas_limit() < gas_after_merge {
-        Err(InvalidPoolTransactionError::IntrinsicGasTooLow)
-    } else {
-        Ok(())
-    }
-}
+//     if transaction.gas_limit() < gas_after_merge {
+//         Err(InvalidPoolTransactionError::IntrinsicGasTooLow)
+//     } else {
+//         Ok(())
+//     }
+// }
 
 #[cfg(test)]
 mod tests {
